feat: decouple account sync code from userstoragecontroller
