+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.ts
@@ -11,13 +11,11 @@ import type {
   StateMetadata,
 } from '@metamask/base-controller';
 import { BaseController } from '@metamask/base-controller';
-import { isEvmAccountType } from '@metamask/keyring-api';
 import {
   type KeyringControllerGetStateAction,
   type KeyringControllerLockEvent,
   type KeyringControllerUnlockEvent,
   type KeyringControllerAddNewAccountAction,
-  KeyringTypes,
 } from '@metamask/keyring-controller';
 import type { InternalAccount } from '@metamask/keyring-internal-api';
 import type {
@@ -32,7 +30,6 @@ import type { HandleSnapRequest } from '@metamask/snaps-controllers';
 import { createSHA256Hash } from '../../shared/encryption';
 import type { UserStorageFeatureKeys } from '../../shared/storage-schema';
 import {
-  USER_STORAGE_FEATURE_NAMES,
   type UserStoragePathWithFeatureAndKey,
   type UserStoragePathWithFeatureOnly,
 } from '../../shared/storage-schema';
@@ -45,11 +42,11 @@ import type {
   AuthenticationControllerPerformSignIn,
   AuthenticationControllerPerformSignOut,
 } from '../authentication/AuthenticationController';
-import type { UserStorageAccount } from './accounts/user-storage';
 import {
-  isNameDefaultAccountName,
-  mapInternalAccountToUserStorageAccount,
-} from './accounts/user-storage';
+  saveInternalAccountToUserStorage,
+  syncInternalAccountsWithUserStorage,
+} from './account-syncing/controller-integration';
+import { setupAccountSyncingSubscriptions } from './account-syncing/setup-subscriptions';
 import {
   performMainNetworkSync,
   startNetworkSyncing,
@@ -98,6 +95,10 @@ export type UserStorageControllerState = {
    */
   isAccountSyncingReadyToBeDispatched: boolean;
   /**
+   * Condition used by UI to determine if account syncing is in progress.
+   */
+  isAccountSyncingInProgress: boolean;
+  /**
    * Condition used to ensure that we do not perform any network sync mutations until we have synced at least once
    */
   hasNetworkSyncingSyncedAtLeastOnce?: boolean;
@@ -108,6 +109,7 @@ export const defaultState: UserStorageControllerState = {
   isProfileSyncingUpdateLoading: false,
   hasAccountSyncingSyncedAtLeastOnce: false,
   isAccountSyncingReadyToBeDispatched: false,
+  isAccountSyncingInProgress: false,
 };
 
 const metadata: StateMetadata<UserStorageControllerState> = {
@@ -124,6 +126,10 @@ const metadata: StateMetadata<UserStorageControllerState> = {
     anonymous: false,
   },
   isAccountSyncingReadyToBeDispatched: {
+    persist: false,
+    anonymous: false,
+  },
+  isAccountSyncingInProgress: {
     persist: false,
     anonymous: false,
   },
@@ -334,121 +340,6 @@ export default class UserStorageController extends BaseController<
     },
   };
 
-  #accounts = {
-    isAccountSyncingInProgress: false,
-    maxNumberOfAccountsToAdd: 0,
-    canSync: () => {
-      try {
-        this.#assertProfileSyncingEnabled();
-
-        if (this.#accounts.isAccountSyncingInProgress) {
-          return false;
-        }
-
-        if (!this.#env.isAccountSyncingEnabled) {
-          return false;
-        }
-
-        if (!this.#auth.isAuthEnabled()) {
-          return false;
-        }
-
-        return true;
-      } catch {
-        return false;
-      }
-    },
-    setupAccountSyncingSubscriptions: () => {
-      this.messagingSystem.subscribe(
-        'AccountsController:accountAdded',
-        // eslint-disable-next-line @typescript-eslint/no-misused-promises
-        async (account) => {
-          if (
-            !this.#accounts.canSync() ||
-            !this.state.hasAccountSyncingSyncedAtLeastOnce
-          ) {
-            return;
-          }
-
-          await this.saveInternalAccountToUserStorage(account);
-        },
-      );
-
-      this.messagingSystem.subscribe(
-        'AccountsController:accountRenamed',
-        // eslint-disable-next-line @typescript-eslint/no-misused-promises
-        async (account) => {
-          if (
-            !this.#accounts.canSync() ||
-            !this.state.hasAccountSyncingSyncedAtLeastOnce
-          ) {
-            return;
-          }
-          await this.saveInternalAccountToUserStorage(account);
-        },
-      );
-    },
-    doesInternalAccountHaveCorrectKeyringType: (account: InternalAccount) => {
-      return account.metadata.keyring.type === KeyringTypes.hd;
-    },
-    getInternalAccountsList: async (): Promise<InternalAccount[]> => {
-      // eslint-disable-next-line @typescript-eslint/await-thenable
-      const internalAccountsList = await this.messagingSystem.call(
-        'AccountsController:listAccounts',
-      );
-
-      return internalAccountsList?.filter(
-        this.#accounts.doesInternalAccountHaveCorrectKeyringType,
-      );
-    },
-    getUserStorageAccountsList: async (): Promise<
-      UserStorageAccount[] | null
-    > => {
-      const rawAccountsListResponse =
-        await this.performGetStorageAllFeatureEntries(
-          USER_STORAGE_FEATURE_NAMES.accounts,
-        );
-
-      return (
-        rawAccountsListResponse?.map((rawAccount) => JSON.parse(rawAccount)) ??
-        null
-      );
-    },
-    saveInternalAccountToUserStorage: async (
-      internalAccount: InternalAccount,
-    ) => {
-      // Map the internal account to the user storage account schema
-      const mappedAccount =
-        mapInternalAccountToUserStorageAccount(internalAccount);
-
-      await this.performSetStorage(
-        // ESLint is confused here.
-        // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
-        `${USER_STORAGE_FEATURE_NAMES.accounts}.${internalAccount.address}`,
-        JSON.stringify(mappedAccount),
-      );
-    },
-    saveInternalAccountsListToUserStorage: async () => {
-      const internalAccountsList =
-        await this.#accounts.getInternalAccountsList();
-
-      if (!internalAccountsList) {
-        return;
-      }
-
-      const internalAccountsListFormattedForUserStorage =
-        internalAccountsList.map(mapInternalAccountToUserStorageAccount);
-
-      await this.performBatchSetStorage(
-        USER_STORAGE_FEATURE_NAMES.accounts,
-        internalAccountsListFormattedForUserStorage.map((account) => [
-          account.a,
-          JSON.stringify(account),
-        ]),
-      );
-    },
-  };
-
   #config?: ControllerConfig;
 
   #notificationServices = {
@@ -516,14 +407,21 @@ export default class UserStorageController extends BaseController<
     this.#env.isNetworkSyncingEnabled = Boolean(env?.isNetworkSyncingEnabled);
     this.#config = config;
 
-    this.#accounts.maxNumberOfAccountsToAdd =
-      config?.accountSyncing?.maxNumberOfAccountsToAdd ?? 100;
-
     this.getMetaMetricsState = getMetaMetricsState;
     this.#keyringController.setupLockedStateSubscriptions();
     this.#registerMessageHandlers();
     this.#nativeScryptCrypto = nativeScryptCrypto;
-    this.#accounts.setupAccountSyncingSubscriptions();
+
+    // Account Syncing
+    if (this.#env.isAccountSyncingEnabled) {
+      setupAccountSyncingSubscriptions(
+        { isAccountSyncingEnabled: true },
+        {
+          getUserStorageControllerInstance: () => this,
+          getMessenger: () => this.messagingSystem,
+        },
+      );
+    }
 
     // Network Syncing
     if (this.#env.isNetworkSyncingEnabled) {
@@ -923,7 +821,7 @@ export default class UserStorageController extends BaseController<
     });
   }
 
-  private async setHasAccountSyncingSyncedAtLeastOnce(
+  async setHasAccountSyncingSyncedAtLeastOnce(
     hasAccountSyncingSyncedAtLeastOnce: boolean,
   ): Promise<void> {
     this.update((state) => {
@@ -932,12 +830,20 @@ export default class UserStorageController extends BaseController<
     });
   }
 
-  public async setIsAccountSyncingReadyToBeDispatched(
+  async setIsAccountSyncingReadyToBeDispatched(
     isAccountSyncingReadyToBeDispatched: boolean,
   ): Promise<void> {
     this.update((state) => {
       state.isAccountSyncingReadyToBeDispatched =
         isAccountSyncingReadyToBeDispatched;
+    });
+  }
+
+  async setIsAccountSyncingInProgress(
+    isAccountSyncingInProgress: boolean,
+  ): Promise<void> {
+    this.update((state) => {
+      state.isAccountSyncingInProgress = isAccountSyncingInProgress;
     });
   }
 
@@ -947,197 +853,28 @@ export default class UserStorageController extends BaseController<
    * It will add new accounts to the internal accounts list, update/merge conflicting names and re-upload the results in some cases to the user storage.
    */
   async syncInternalAccountsWithUserStorage(): Promise<void> {
-    if (!this.#accounts.canSync()) {
-      return;
-    }
+    const profileId = await this.#auth.getProfileId();
 
-    try {
-      this.#accounts.isAccountSyncingInProgress = true;
-
-      const profileId = await this.#auth.getProfileId();
-
-      const userStorageAccountsList =
-        await this.#accounts.getUserStorageAccountsList();
-
-      if (!userStorageAccountsList || !userStorageAccountsList.length) {
-        await this.#accounts.saveInternalAccountsListToUserStorage();
-        await this.setHasAccountSyncingSyncedAtLeastOnce(true);
-        return;
-      }
-
-      // Prepare an array of internal accounts to be saved to the user storage
-      const internalAccountsToBeSavedToUserStorage: InternalAccount[] = [];
-
-      // Compare internal accounts list with user storage accounts list
-      // First step: compare lengths
-      const internalAccountsList =
-        await this.#accounts.getInternalAccountsList();
-
-      if (!internalAccountsList || !internalAccountsList.length) {
-        throw new Error(`Failed to get internal accounts list`);
-      }
-
-      const hasMoreUserStorageAccountsThanInternalAccounts =
-        userStorageAccountsList.length > internalAccountsList.length;
-
-      // We don't want to remove existing accounts for a user
-      // so we only add new accounts if the user has more accounts in user storage than internal accounts
-      if (hasMoreUserStorageAccountsThanInternalAccounts) {
-        const numberOfAccountsToAdd =
-          Math.min(
-            userStorageAccountsList.length,
-            this.#accounts.maxNumberOfAccountsToAdd,
-          ) - internalAccountsList.length;
-
-        // Create new accounts to match the user storage accounts list
-        for (let i = 0; i < numberOfAccountsToAdd; i++) {
-          await this.messagingSystem.call('KeyringController:addNewAccount');
-          this.#config?.accountSyncing?.onAccountAdded?.(profileId);
-        }
-      }
-
-      // Second step: compare account names
-      // Get the internal accounts list again since new accounts might have been added in the previous step
-      const refreshedInternalAccountsList =
-        await this.#accounts.getInternalAccountsList();
-
-      const newlyAddedAccounts = refreshedInternalAccountsList.filter(
-        (account) =>
-          !internalAccountsList.find((a) => a.address === account.address),
-      );
-
-      for (const internalAccount of refreshedInternalAccountsList) {
-        const userStorageAccount = userStorageAccountsList.find(
-          (account) => account.a === internalAccount.address,
-        );
-
-        // If the account is not present in user storage
-        if (!userStorageAccount) {
-          // If the account was just added in the previous step, skip saving it, it's likely to be a bogus account
-          if (newlyAddedAccounts.includes(internalAccount)) {
-            this.#config?.accountSyncing?.onAccountSyncErroneousSituation?.(
-              profileId,
-              'An account was added to the internal accounts list but was not present in the user storage accounts list',
-            );
-            continue;
-          }
-          // Otherwise, it means that this internal account was present before the sync, and needs to be saved to the user storage
-          internalAccountsToBeSavedToUserStorage.push(internalAccount);
-          continue;
-        }
-
-        // From this point on, we know that the account is present in
-        // both the internal accounts list and the user storage accounts list
-
-        // One or both accounts have default names
-        const isInternalAccountNameDefault = isNameDefaultAccountName(
-          internalAccount.metadata.name,
-        );
-        const isUserStorageAccountNameDefault = isNameDefaultAccountName(
-          userStorageAccount.n,
-        );
-
-        // Internal account has default name
-        if (isInternalAccountNameDefault) {
-          if (!isUserStorageAccountNameDefault) {
-            this.messagingSystem.call(
-              'AccountsController:updateAccountMetadata',
-              internalAccount.id,
-              {
-                name: userStorageAccount.n,
-              },
-            );
-
-            this.#config?.accountSyncing?.onAccountNameUpdated?.(profileId);
-          }
-          continue;
-        }
-
-        // Internal account has custom name but user storage account has default name
-        if (isUserStorageAccountNameDefault) {
-          internalAccountsToBeSavedToUserStorage.push(internalAccount);
-          continue;
-        }
-
-        // Both accounts have custom names
-
-        // User storage account has a nameLastUpdatedAt timestamp
-        // Note: not storing the undefined checks in constants to act as a type guard
-        if (userStorageAccount.nlu !== undefined) {
-          if (internalAccount.metadata.nameLastUpdatedAt !== undefined) {
-            const isInternalAccountNameNewer =
-              internalAccount.metadata.nameLastUpdatedAt >
-              userStorageAccount.nlu;
-
-            if (isInternalAccountNameNewer) {
-              internalAccountsToBeSavedToUserStorage.push(internalAccount);
-              continue;
-            }
-          }
-
-          this.messagingSystem.call(
-            'AccountsController:updateAccountMetadata',
-            internalAccount.id,
-            {
-              name: userStorageAccount.n,
-              nameLastUpdatedAt: userStorageAccount.nlu,
-            },
-          );
-
-          const areInternalAndUserStorageAccountNamesEqual =
-            internalAccount.metadata.name === userStorageAccount.n;
-
-          if (!areInternalAndUserStorageAccountNamesEqual) {
-            this.#config?.accountSyncing?.onAccountNameUpdated?.(profileId);
-          }
-
-          continue;
-        } else if (internalAccount.metadata.nameLastUpdatedAt !== undefined) {
-          internalAccountsToBeSavedToUserStorage.push(internalAccount);
-          continue;
-        }
-      }
-
-      // Save the internal accounts list to the user storage
-      if (internalAccountsToBeSavedToUserStorage.length) {
-        await this.performBatchSetStorage(
-          USER_STORAGE_FEATURE_NAMES.accounts,
-          internalAccountsToBeSavedToUserStorage.map((account) => [
-            account.address,
-            JSON.stringify(mapInternalAccountToUserStorageAccount(account)),
-          ]),
-        );
-      }
-
-      // In case we have corrupted user storage with accounts that don't exist in the internal accounts list
-      // Delete those accounts from the user storage
-      const userStorageAccountsToBeDeleted = userStorageAccountsList.filter(
-        (account) =>
-          !refreshedInternalAccountsList.find((a) => a.address === account.a),
-      );
-
-      if (userStorageAccountsToBeDeleted.length) {
-        await this.performBatchDeleteStorage(
-          USER_STORAGE_FEATURE_NAMES.accounts,
-          userStorageAccountsToBeDeleted.map((account) => account.a),
-        );
-        this.#config?.accountSyncing?.onAccountSyncErroneousSituation?.(
-          profileId,
-          'An account was present in the user storage accounts list but was not found in the internal accounts list after the sync',
-        );
-      }
-
-      // We do this here and not in the finally statement because we want to make sure that
-      // the accounts are saved / updated / deleted at least once before we set this flag
-      await this.setHasAccountSyncingSyncedAtLeastOnce(true);
-    } catch (e) {
-      const errorMessage = e instanceof Error ? e.message : JSON.stringify(e);
-      throw new Error(
-        `${controllerName} - failed to sync user storage accounts list - ${errorMessage}`,
-      );
-    } finally {
-      this.#accounts.isAccountSyncingInProgress = false;
-    }
+    await syncInternalAccountsWithUserStorage(
+      {
+        isAccountSyncingEnabled: this.#env.isAccountSyncingEnabled,
+        maxNumberOfAccountsToAdd:
+          this.#config?.accountSyncing?.maxNumberOfAccountsToAdd,
+        onAccountAdded: () =>
+          this.#config?.accountSyncing?.onAccountAdded?.(profileId),
+        onAccountNameUpdated: () =>
+          this.#config?.accountSyncing?.onAccountNameUpdated?.(profileId),
+        onAccountSyncErroneousSituation: (situationMessage) =>
+          this.#config?.accountSyncing?.onAccountSyncErroneousSituation?.(
+            profileId,
+            situationMessage,
+          ),
+      },
+      {
+        getMessenger: () => this.messagingSystem,
+        getUserStorageControllerInstance: () => this,
+      },
+    );
   }
 
   /**
@@ -1147,22 +884,14 @@ export default class UserStorageController extends BaseController<
   async saveInternalAccountToUserStorage(
     internalAccount: InternalAccount,
   ): Promise<void> {
-    if (
-      !this.#accounts.canSync() ||
-      !isEvmAccountType(internalAccount.type) ||
-      !this.#accounts.doesInternalAccountHaveCorrectKeyringType(internalAccount)
-    ) {
-      return;
-    }
-
-    try {
-      await this.#accounts.saveInternalAccountToUserStorage(internalAccount);
-    } catch (e) {
-      const errorMessage = e instanceof Error ? e.message : JSON.stringify(e);
-      throw new Error(
-        `${controllerName} - failed to save account to user storage - ${errorMessage}`,
-      );
-    }
+    await saveInternalAccountToUserStorage(
+      internalAccount,
+      { isAccountSyncingEnabled: this.#env.isAccountSyncingEnabled },
+      {
+        getMessenger: () => this.messagingSystem,
+        getUserStorageControllerInstance: () => this,
+      },
+    );
   }
 
   async syncNetworks() {
