+++ packages/profile-sync-controller/src/controllers/user-storage/__fixtures__/test-utils.ts
@@ -1,3 +1,65 @@
+import type nock from 'nock';
+
+import encryption from '../../../shared/encryption/encryption';
+import type {
+  GetUserStorageAllFeatureEntriesResponse,
+  GetUserStorageResponse,
+} from '../services';
+import { MOCK_STORAGE_KEY } from './mockStorage';
+
+/**
+ * Test Utility - creates a realistic mock user-storage entry
+ * @param data - data to encrypt
+ * @returns user storage entry
+ */
+export async function createMockUserStorageEntry(
+  data: unknown,
+): Promise<GetUserStorageResponse> {
+  return {
+    HashedKey: 'HASHED_KEY',
+    Data: await encryption.encryptString(
+      JSON.stringify(data),
+      MOCK_STORAGE_KEY,
+    ),
+  };
+}
+
+/**
+ * Test Utility - creates a realistic mock user-storage get-all entry
+ * @param data - data array to encrypt
+ * @returns user storage entry
+ */
+export async function createMockUserStorageEntries(
+  data: unknown[],
+): Promise<GetUserStorageAllFeatureEntriesResponse> {
+  return await Promise.all(data.map((d) => createMockUserStorageEntry(d)));
+}
+
+/**
+ * Test Utility - decrypts a realistic batch upsert payload
+ * @param requestBody - nock body
+ * @param storageKey - storage key
+ * @returns decrypted body
+ */
+export async function decryptBatchUpsertBody(
+  requestBody: nock.Body,
+  storageKey: string,
+) {
+  if (typeof requestBody === 'string') {
+    return requestBody;
+  }
+  return await Promise.all(
+    Object.entries<string>(requestBody.data).map(
+      async ([entryKey, entryValue]) => {
+        return [
+          entryKey,
+          await encryption.decryptString(entryValue, storageKey),
+        ];
+      },
+    ),
+  );
+}
+
 type WaitForOptions = {
   intervalMs?: number;
   timeoutMs?: number;
